[refactor] : GuestList 컴포넌트에서 직접 Customers.json 데이터를 불러오는 것으로 변경
